// src/shared/api/generated/__swagger_client/models/ReturnUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ReturnLineCreateRequest } from './ReturnLineCreateRequest';
export type ReturnUpdateRequest = {
    docNumber?: string;
    docDate?: string;
    organizationId?: string;
    warehouseId?: string;
    employeeId?: string;
    employeeFullName?: string;
    reason?: string;
    lines?: Array<ReturnLineCreateRequest>;
};